Fix missing `key` prop handling in HKModal
Because we dynamically include some elements on the page, React
appears to get confused in some situations (have not been able to
reproduce in Storybook, but repros easily in DHC whenever the modal
for rotate credentials is opened) as to which elements are which and
insists on keys. We don't want to modify passed-in elements to set key
props on them, so instead we wrap the header element in a div, which
appears to work around the issue.

Fixes #56