Drop "break lodash" plugin
It breaks lodash.

We were not quite able to figure out what's going on, but apparently
this plugin, while it does improve our bundle size slightly, also
makes a total mess of (some?) lodash methods, which led to very
strange behavior when RHKC was used in DHC (and probably elsewhere).

Remove the plugin, and rely on the built-in Array.flatMap since that's
simpler (we should only rely on lodash where necessary). Also some
minor refactoring in related code, and moves lodash from
devDependencies where we mistakenly had it to a proper dependency.