Fix charting lodash shenanigans (#64)
Drop the "break lodash" plugin. It breaks lodash.

We were not quite able to figure out what's going on, but apparently
this plugin, while it does improve our bundle size slightly, also
makes a total mess of (some?) lodash methods, which led to very
strange behavior when RHKC was used in DHC (and probably elsewhere).

After removing the plugin, we noticed we could rely on browsers' built-in
Array.prototype.flatMap, but it turns out node does not define that so it
broke our tests, so we'll keep using lodash.

Also some minor refactoring in related code, and move lodash from
devDependencies where we mistakenly had it to a proper dependency.